--- a/app/src/main/java/com/bs/threadsimulator/di/AppModule.kt
+++ b/app/src/main/java/com/bs/threadsimulator/di/AppModule.kt
@@ -1,0 +1,19 @@
+package com.bs.threadsimulator.di
+
+import com.bs.threadsimulator.common.ThreadMonitor
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object AppModule {
+    
+    @Provides
+    @Singleton
+    fun provideThreadMonitor(): ThreadMonitor {
+        return ThreadMonitor()
+    }
+}