--- a/app/src/main/java/com/bs/threadsimulator/ui/screens/components/CompanyItem.kt
+++ b/app/src/main/java/com/bs/threadsimulator/ui/screens/components/CompanyItem.kt
@@ -9,6 +9,10 @@ import androidx.compose.material3.ElevatedCard
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.derivedStateOf
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberUpdatedState
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.colorResource
@@ -20,10 +24,15 @@ import com.bs.threadsimulator.R
 import com.bs.threadsimulator.model.Company
 
 @Composable
-fun CompanyItem(company: Company) {
+fun CompanyItem(
+    company: Company,
+    modifier: Modifier
+) {
+    val currentPrice by rememberUpdatedState(company.stock.currentPrice)
+
     ElevatedCard(
         colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surfaceVariant),
-        modifier = Modifier
+        modifier = modifier
             .padding(horizontal = 8.dp)
             .fillMaxWidth()
             .clickable(enabled = true, onClick = {}),
@@ -53,12 +62,20 @@ fun CompanyItem(company: Company) {
             text = "Closing Price: ${company.previousClosingPrice}",
             modifier = Modifier.padding(horizontal = 8.dp, vertical = 4.dp)
         )
+
+        // Use remember for expensive computations
+        val priceColor by remember(currentPrice) {
+            derivedStateOf {
+                if (currentPrice > company.stock.openingPrice) Color(0xFF2BAB2B) else Color.Red
+            }
+        }
+
         Text(
             text = "Price: ${company.stock.currentPrice}",
             modifier = Modifier.padding(horizontal = 8.dp, vertical = 4.dp),
             fontWeight = FontWeight.Bold,
             fontSize = 22.sp,
-            color = colorResource(R.color.purple_200)
+            color = priceColor,
         )
         Row {
             Text(