**Key Improvements:**

1. Added thread monitoring to visualize how updates are distributed across threads

2. Optimized LazyList performance with proper keys and contentType

3. Used rememberUpdatedState and derivedStateOf to minimize recompositions

4. Added throttling mechanism to prevent too frequent updates

5. Kept the multi-threaded update simulation while adding metrics

6. Used proper Composable optimization techniques

**The benefits of this implementation:**

**1. Prevents UI Overload:** 
• 16ms throttle = ~60 FPS (very smooth)
• 32ms throttle = ~30 FPS (good balance)
• 64ms throttle = ~15 FPS (more relaxed updates)

**3. Flexible Configuration:** 
• You can set different throttle rates for different types of updates 
• Can be changed dynamically based on performance needs

**5. Resource Efficiency:** 
• Reduces number of recompositions in your LazyList 
• Lower CPU usage due to fewer updates
• Better battery life in real applications

**6. Testing Different Scenarios:** 
• You can easily test how your UI performs under different update frequencies 
• Helps identify the optimal update rate for your use case